+++ archive.php
@@ -17,9 +17,10 @@ get_header(); ?>
 
         <header class="page-header">
             <?php
-            // Standart solution for this type of task
-            //the_archive_title( '<h1 class="page-title">', '</h1>' );
-            //the_archive_description( '<div class="taxonomy-description">', '</div>' );
+            /* Standart solution for this type of task:
+            * - the_archive_title( '<h1 class="page-title">', '</h1>' );
+            * - the_archive_description( '<div class="taxonomy-description">', '</div>' );
+            */
             ?>
 
             <?php