+++ widgets/artfolio-recent-posts.php
@@ -1,3 +1,10 @@
+/**
+ * Custom widget with recent posts links, images and meta-info
+ *
+ * @package Artfolio
+ */
+
+
 <?php
 
 // Widget for recent posts with image, title, date and author
@@ -18,45 +25,45 @@ class artfolio_recent_posts extends WP_Widget {
 
     // Open code for widget
     public function widget( $args, $instance ) {
-        $title = apply_filters( 'widget_title', $instance['title'] );
-        $number = apply_filters( 'widget_number', absint( $instance['number'] ) );
+        $artfolio_title = apply_filters( 'widget_title', $instance['title'] );
+        $artfolio_number = apply_filters( 'widget_number', absint( $instance['number'] ) );
 
         echo $args['before_widget'];
 
-        if ( !empty( $title ) ) {
-            echo $args['before_title'] . $title . $args['after_title'];
+        if ( !empty( $artfolio_title ) ) {
+            echo $args['before_title'] . $artfolio_title . $args['after_title'];
         }
 
-        $art_recent_posts = new WP_Query();
-        $art_recent_posts->query( 'showposts=' . $number );
+        $artfolio_recent_posts = new WP_Query();
+        $artfolio_recent_posts->query( 'showposts=' . $artfolio_number );
 
-        while ( $art_recent_posts->have_posts() ) {
-            $art_recent_posts->the_post();
+        while ( $artfolio_recent_posts->have_posts() ) {
+            $artfolio_recent_posts->the_post();
 ?>
 <ul>
     <li>
 
         <?php if ( has_post_thumbnail() ) { ?>
-        <a href="<?php the_permalink(); ?>" class="arp-thumbnail">
-            <?php the_post_thumbnail( 'artfolio-recent-thumbnails' ); ?>
+        <a href="<?php the_permalink(); ?>" class="artfolio-posts-widget-thumbnail">
+            <?php the_post_thumbnail( 'artfolio-posts-widget-thumbnails' ); ?>
         </a>
         <?php } else {?>
-        <a href="<?php the_permalink(); ?>" class="arp-thumbnail imagePlugWrapper">
-            <i class="fa fa-picture-o imagePlug" aria-hidden="true"></i>
+        <a href="<?php the_permalink(); ?>" class="artfolio-posts-widget-thumbnail artfolio-posts-widget-image-plug-wrapper">
+            <i class="fa fa-picture-o artfolio-posts-widget-image-plug" aria-hidden="true"></i>
         </a>
         <?php }?>
 
-        <header class="arp-header"
+        <header class="artfolio-posts-widget-header"
                 <h3>
         <a href="<?php the_permalink(); ?>" class="arp-title"><?php esc_html( the_title() ); ?></a>
     </h3>
 
-<div class="arp-meta">
-    <p class="arp-time"><?php the_time( 'F jS, Y' ); ?></p>
-    <a href="<?php echo get_author_posts_url( get_the_author_meta( 'ID' ) ); ?>" class="arp-author">
+<div class="artfolio-posts-widget-meta">
+    <p class="artfolio-posts-widget-time"><?php the_time( 'F jS, Y' ); ?></p>
+    <a href="<?php echo get_author_posts_url( get_the_author_meta( 'ID' ) ); ?>" class="artfolio-posts-widget-author">
         <?php esc_html( the_author() ); ?>
     </a>
-</div> <!-- .arp-meta-->
+</div> <!-- .artfolio-posts-widget-meta-->
 <?php
             /* translators: used between list items, there is a space after the comma */
             $category_list = get_the_category_list( __( ', ', 'artfolio' ) );
@@ -65,7 +72,7 @@ class artfolio_recent_posts extends WP_Widget {
                 echo '<div class="category-list">' . $category_list . '</div>';
             }
 ?>
-</header> <!-- .arp-header-->
+</header> <!-- .artfolio-posts-widget-header-->
 </li>
 </ul>
 <?php
@@ -83,27 +90,27 @@ class artfolio_recent_posts extends WP_Widget {
     public function form( $instance ) {
 
         if ( isset( $instance[ 'number' ] ) ) {
-            $number = $instance[ 'number' ];
+            $artfolio_number = $instance[ 'number' ];
         } else {
-            $number = 5;
+            $artfolio_number = 5;
         }
 
         if ( isset( $instance[ 'title' ] ) ) {
-            $title = $instance[ 'title' ];
+            $artfolio_title = $instance[ 'title' ];
         } else {
-            $title = __( 'Recent posts', 'artfolio' );
+            $artfolio_title = __( 'Recent posts', 'artfolio' );
         }
 
         // For administration console
 ?>
 <p>
     <label for="<?php echo $this->get_field_id( 'title' ); ?>"><?php _e( 'Title:', 'artfolio' ); ?></label> 
-    <input class="widefat" id="<?php echo $this->get_field_id( 'title' ); ?>" name="<?php echo $this->get_field_name( 'title' ); ?>" type="text" value="<?php echo esc_attr( $title ); ?>" />
+    <input class="widefat" id="<?php echo $this->get_field_id( 'title' ); ?>" name="<?php echo $this->get_field_name( 'title' ); ?>" type="text" value="<?php echo esc_attr( $artfolio_title ); ?>" />
 </p>
 
 <p>
     <label for="<?php echo $this->get_field_id( 'number' ); ?>"><?php _e( 'Number of posts to show:', 'artfolio' ); ?></label>
-    <input class="tiny-text" id="<?php echo $this->get_field_id( 'number' ); ?>" name="<?php echo $this->get_field_name( 'number' ); ?>" type="number" step="1" min="1" value="<?php echo $number; ?>" size="3" />
+    <input class="tiny-text" id="<?php echo $this->get_field_id( 'number' ); ?>" name="<?php echo $this->get_field_name( 'number' ); ?>" type="number" step="1" min="1" value="<?php echo $artfolio_number; ?>" size="3" />
 </p>
 
 <?php